repsense-AI/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
websockets==12.0
google-generativeai==0.7.0
ultralytics==8.1.0
mediapipe==0.10.11
opencv-python-headless==4.9.0.80
numpy==1.26.4
python-multipart==0.0.9
pillow==10.3.0
pyttsx3==2.90
